component: SpaceWeekCard에 cardList prop 추가 및 Body 컴포넌트 구현
