+++ src/app/page.tsx
@@ -6,6 +6,7 @@ import { useState } from "react";
 import TitleHeading from "./_components/common/heding/TitleHeading";
 import { 시간변수 } from "./_components/search/constant";
 import Space from "./_components/space/Space";
+import { cardSectionList } from "./_components/space/SpaceConstant";
 import SpaceDetailHead from "./_components/space/SpaceDetailHead";
 import SpaceWeekCard from "./_components/space/SpaceWeekCard";
 import CalendarUI from "./_components/space/WriteCalendar";
@@ -38,7 +39,7 @@ export default function Home() {
 				minYear={2021}
 				maxYear={2022}
 			/>
-			<SpaceWeekCard />
+			<SpaceWeekCard cardList={cardSectionList} />
 			<WriteForm />
 
 			<button onClick={openActive}>모달 열기</button>
